requirements.txt
flask
fastapi[all]
mysql-connector-python